# Significantly improve performance of row decoding.