@mcdappdev 2019-03-07
Can this be applied to MySQL as well or is it Postgres specific?