@MrMage 2019-03-07
@mcdappdev looks like https://github.com/vapor/mysql/blob/master/Sources/MySQL/Codable/MySQLRowDecoder.swift is doing a linear search as well (`firstValue(forColumn:`), so Optimization 1. Not sure if a table name cache is being used in the MySQL driver, so Optimization 2 might not apply.