@mcdappdev 2019-03-07
Cool... @tanner0101 thoughts?