@twof 2019-03-07
@MrMage Are you able to include the performance tests you're running?

Great work btw!