@MrMage 2019-03-08
> @MrMage Are you able to include the performance tests you're running?

My tests were run with production code and data, so I'm afraid I can't share them. Also, I'm not sure whether Fluent has a good way to observe and record performance improvements/regressions, anyway. (There is https://github.com/vapor/fluent/tree/3/Sources/FluentBenchmark/Benchmarks, but I don't see how changes in these benchmarks' performance would be exposed and recorded.)