@MrMage 2019-03-08
Added another micro-optimization (implementing `decodeIfPresent` to avoid two dictionary lookups). Further gains _might_ be had from overriding all the methods in `KeyedDecodingContainerProtocol` for the types which have specializations (`Bool`, `String`, `Double`, `Int`, etc.), provided `PostgreSQLDataDecoder.decoded` is being specialized in the same fashion.