@twof 2019-03-08
@MrMage Working on something right now to track performance metrics
https://github.com/vapor/bot-github
https://github.com/vapor/routing/pull/66#issuecomment-466966608

Hoping to have the bot redeployed tomorrow and then integrated with this repo so we can get you some charts :)