@MrMage 2019-03-11
@twof sounds good; let me know how to use that bot when its done (and where to add tests to showcase the performance gains; they would be most apparent for DBs with many tables and/or tables with many columns).