@tanner0101 2019-03-11
> they would be most apparent for DBs with many tables and/or tables with many columns

We should add some performance tests for specifically this use case. Most of them end up being against dummy models with very few columns which gives unrealistic results.

Generally, this should be less of a problem in Fluent 4 since we will no longer be decoding all model properties by default. Properties will be decoded lazily on an as needed basis. (i.e., `user.name.get()`). However, it might be worth optimizing PostgresRow's value fetch method. Maybe we can use a dictionary here? Here's the relevant code in NIOPostgres which we will be using for PostgresKit 2.0: 

https://github.com/vapor/nio-postgres/blob/master/Sources/NIOPostgres/Data/PostgresRow.swift#L16