@tanner0101 2019-03-12
Postgres sends back [one `RowDescription`](https://github.com/vapor/nio-postgres/blob/master/Sources/NIOPostgres/Connection/PostgresDatabase%2BQuery.swift#L46-L51) per query. That's why I made [`PostgresRow.LookupTable`](https://github.com/vapor/nio-postgres/blob/master/Sources/NIOPostgres/Data/PostgresRow.swift#L2) a class type. That way we can share it between rows without worrying about copying. 

So we should be able to create a lookup dictionary once (we can even do it lazily), then re-use for _all_ rows. That should be an even bigger performance win than is possible in the current version.