@MrMage 2019-03-12
Sounds great! Is there any possibility of backporting the re-use of the column description to Fluent 3 as well?

> On 12. Mar 2019, at 19:14, Tanner <notifications@github.com> wrote:
> 
> Postgres sends back one RowDescription per query. That's why I made PostgresRow.LookupTable a class type. That way we can share it between rows without worrying about copying.
> 
> So we should be able to create a lookup dictionary once (we can even do it lazily), then re-use for all rows. That should be an even bigger performance win than is possible in the current version.
> 
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub, or mute the thread.