@tanner0101 2019-03-12
I'm not sure, it might be.

Is this PR ready to merge? I've reverted #117 so the `1` branch should be good to go.