@MrMage 2019-03-12
This PR is ready to merge; I had just been waiting for better benchmark infrastructure to demonstrate the improvements with hard numbers (and possibly add a benchmark case for this first). It the benchmark infrastructure is already working, please let me know (a pointer to the instructions would be app).

> On 12. Mar 2019, at 19:55, Tanner <notifications@github.com> wrote:
> 
> I'm not sure, it might be.
> 
> Is this PR ready to merge? I've reverted #117 so the 1 branch should be good to go.
> 
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub, or mute the thread.