@tanner0101 2019-03-12
I looked into adding the performance testing bot to `1` and it looks like it won't be possible without a significant amount of work. `@testable` is not usable with `swift build -c release` on Linux for some reason, and a lot of our tests use internal types. We'll need to find a way around this. But I don't want that to hold up this PR, so I think we should go ahead and merge if you all agree.