@MrMage 2019-03-12
I'm fine with merging now; however we still have two options left that we could discuss first:

1. Run benchmarks in Debug mode. This won't give us 100% accurate data, but trends and changes in algorithmic complexity (eg from quadraric to linear should become apparent).
2. Run benchmarks in Release mode, as part of a separate target. We should be able to create realistic testing scenarios without having to rely on internal APIs, after all.

> On 12. Mar 2019, at 21:53, Tanner <notifications@github.com> wrote:
> 
> I looked into adding the performance testing bot to 1 and it looks like it won't be possible without a significant amount of work. @testable is not usable with swift build -c release on Linux for some reason, and a lot of our tests use internal types. We'll need to find a way around this. But I don't want that to hold up this PR, so I think we should go ahead and merge if you all agree.
> 
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub, or mute the thread.