@tanner0101 2019-03-13
> Run benchmarks in Release mode, as part of a separate target. We should be able to create realistic testing scenarios without having to rely on internal APIs, after all.

I like this option, but I'm not sure how we'd do it. Any use of `@testable` causes build errors when using `swift test -c release`, and AFAIK you must compile the entire package with each invocation of `swift build` or `swift test`. 

I think the real solution here is to make `@testable` work in release mode on Linux.