@MrMage 2019-03-13
> I like this option, but I'm not sure how we'd do it. Any use of `@testable` causes build errors when using `swift test -c release`, and AFAIK you must compile the entire package with each invocation of `swift build` or `swift test`.
> 
> I think the real solution here is to make `@testable` work in release mode on Linux.

If meant creating "benchmarks" target with _different_ tests that don't use any `@testable` APIs. How about that?