@tanner0101 2019-03-13
@MrMage see my edit, I think `swift test -c release -Xswiftc -enable-testing` might work