@tanner0101 2019-03-13
Our bot is built on `XCTest.measure` so it would be a significant amount of work to start doing testing in executable targets.