@penny-coin 2019-03-13
Hey @MrMage, you just merged a pull request, have a coin! 

You now have 13 coins.