@MrMage 2019-03-13
> Our bot is built on `XCTest.measure` so it would be a significant amount of work to start doing testing in executable targets.

But couldn't we use `XCTest.measure` _without_ `@testable`? I.e. _have an extra "benchmarks" test target with custom benchmark code that does not use `@testable`_? (Trying `-Xswiftc -enable-testing` is also an option, of course, but I feel like we should be able to get by without `@testable`, anyway.